+++ app/Http/Controllers/User/ProfileController.php
@@ -32,35 +32,38 @@ class ProfileController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \App\Http\Requests\ProfileRequest  $request
-     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(ProfileRequest $request, $id)
+    public function update(ProfileRequest $request)
     {
         $data = $request->all();
 
-        $user = User::find($id);
+        $user = auth()->user();
 
-        $oldImageProfile = $user->profile_image;
-
-        if (!is_null($request->profile_image))
+        if (!is_null($user))
         {
-            Storage::delete('public/' . $oldImageProfile);
-            $profileImage = $request->profile_image->store('img', ['disk' => 'public']);
+            $oldImageProfile = $user->profile_image;
 
-            $data['profile_image'] = $profileImage;
+            if (!is_null($request->profile_image))
+            {
+                Storage::delete('public/' . $oldImageProfile);
+                $profileImage = $request->profile_image->store('img', ['disk' => 'public']);
 
-        } else {
-            $data['profile_image'] = $oldImageProfile;
+                $data['profile_image'] = $profileImage;
+
+            } else {
+                $data['profile_image'] = $oldImageProfile;
+            }
+
+            $user->update(
+                [
+                    'profile_image' => $data['profile_image'],
+                ]
+            );
+
+            flash('success', __('Profile updated successfully'));
+
         }
-
-        $user->update(
-            [
-                'profile_image' => $data['profile_image'],
-            ]
-        );
-
-        flash('success', __('Profile updated successfully.'));
 
         return redirect(route('user.profile.show'));
